Code snippet that highlights the interaction system used in my Unity games
This feature implements the Interaction class which handles the various interactions that a Player can have with the different Interactable objects and characters in a game.

Player gameobject will require an Interaction Point which sits ahead of the Player gameObject's transform and  collides as a trigger.

It should display an Interaction prompt with associate hotkey as well as text indicating what kind of interaction it is.